requirements.txt
dash
mysql-connector-python
pymongo
neo4j
python-dotenv
wordcloud
networkx
